+++ Sources/Sentry/PrivateSentrySDKOnly.m
@@ -284,9 +284,8 @@ static BOOL _framesTrackingMeasurementHybridSDKMode = NO;
 #if SENTRY_TARGET_REPLAY_SUPPORTED
     // As the options are not passed in by the hybrid SDK, we need to use the options from the
     // current hub.
-    SentryViewScreenshotOptions *_Nonnull options = PrivateSentrySDKOnly.options.screenshot;
-    SentryScreenshotSource *_Nonnull screenshotSource =
-        [SentryDependencyContainer.sharedInstance getScreenshotSourceForOptions:options];
+    SentryScreenshotSource *_Nonnull screenshotSource
+        = SentryDependencyContainer.sharedInstance.screenshotSource;
     return [screenshotSource appScreenshotsData];
 #else
     SENTRY_LOG_DEBUG(